Try to "guess" the favicon when not parsed
Favicons can often be found at `<site_base_url>/favicon.ico`, even when
they cannot be parsed from the site HTML. Instead of just returning
the base URL in that instance, append the `/favicon.ico` to the base
URL in an attempt to guess its location. So far I haven't found a
site that this hasn't worked for (previously about 50% of sites would
not have a favicon displayed).